requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.25.0
confusable-homoglyphs==3.2.0
Django==3.2.3
django-bootstrap-v5==1.0.4
django-registration==3.1.2
djangorestframework==3.12.4
importlib-metadata==3.10.1
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
zipp==3.4.1
